+++ vue/components/ui/atoms/tag/tag.vue
@@ -139,6 +139,10 @@ export const Tag = {
             type: String,
             default: null
         },
+        textColor: {
+            type: String,
+            default: null
+        },
         subtle: {
             type: Boolean,
             default: true
@@ -155,6 +159,7 @@ export const Tag = {
         style() {
             const base = {};
             if (this.colorHex) base["background-color"] = this.colorHex;
+            if (this.textColor) base.color = this.textColor;
             return base;
         }
     },